pyproject.toml
[project]
name = "spotify"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.11",
    "ipython>=9.0.2",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "spotipy>=2.25.1",
]
